requirements.txt
# -*- coding: utf-8 -*-
# ===== 核心框架 =====
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6  # 文件上传支持

# ===== 数据验证 =====
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0  # 邮箱验证

# ===== LangChain 生态 =====
langgraph>=0.0.20
langchain>=0.1.0
langchain-openai>=0.0.1
langchain-community>=0.0.10

# ===== 数据库 =====
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0  # PostgreSQL 异步驱动
alembic==1.12.1  # 数据库迁移

# ===== 向量数据库 =====
chromadb>=1.0.0,<2.0.0  # 向量数据库（兼容 Chroma v2 API）
pgvector==0.2.4  # PostgreSQL 向量扩展（可选）

# ===== Redis 缓存 =====
redis==5.0.1
hiredis==2.2.3  # Redis 性能优化

# ===== HTTP 客户端 =====
httpx>=0.27.0,<0.29.0  # 异步 HTTP 客户端（与 chromadb 兼容）
aiohttp==3.9.1  # 异步 HTTP

# ===== 文档处理 =====
pypdf==3.17.4  # PDF 解析
python-docx==1.1.0  # Word 解析
python-pptx==0.6.23  # PowerPoint 解析
openpyxl==3.1.2  # Excel 解析
beautifulsoup4==4.12.2  # HTML 解析
markdownify==0.11.6  # HTML 转 Markdown

# ===== 文本处理 =====
jieba==0.42.1  # 中文分词
rank-bm25==0.2.2  # BM25 算法
numpy==1.26.2  # 数值计算

# ===== 认证与安全 =====
python-jose[cryptography]==3.3.0  # JWT 处理
passlib[bcrypt]==1.7.4  # 密码加密
bcrypt==4.0.1  # 与 passlib 1.7.4 兼容的 bcrypt 版本
python-multipart==0.0.6  # 表单数据

# ===== 限流与防护 =====
slowapi==0.1.9  # 限流
itsdangerous==2.1.2  # 签名

# ===== 任务队列 =====
celery==5.3.4  # 异步任务队列
redis==5.0.1  # Celery broker

# ===== 监控与日志 =====
loguru==0.7.2  # 日志
prometheus-client==0.19.0  # 监控指标

# ===== 工具库 =====
python-dotenv==1.0.0  # 环境变量
tenacity==8.2.3  # 重试机制
python-dateutil==2.8.2  # 日期处理

# ===== 开发工具 =====
pytest==7.4.3
pytest-asyncio==0.21.1
httpx>=0.27.0,<0.29.0  # 测试客户端
black==23.12.1  # 代码格式化
flake8==6.1.0  # 代码检查
mypy==1.7.1  # 类型检查

# ===== 可选：Embedding 模型 =====
# sentence-transformers==2.2.2  # 本地 embedding
# cohere==4.39  # Rerank API
# openai==1.3.7  # OpenAI API
